+++ src/components/Table/hooks/useTableSort.js
@@ -1,11 +1,14 @@
+import { camelToSnakeCase } from "neetocist";
 import { mergeLeft } from "ramda";
 import { useHistory } from "react-router-dom";
 
+import { useQueryParams } from "hooks";
+import { buildUrl } from "utils";
+
 import { URL_SORT_ORDERS } from "../constants";
-import { camelToSnakeCase, getQueryParams, buildUrl } from "../utils";
 
 const useTableSort = () => {
-  const queryParams = getQueryParams();
+  const queryParams = useQueryParams();
 
   const history = useHistory();
 
@@ -14,7 +17,6 @@ const useTableSort = () => {
       sort_by: sorter.order && camelToSnakeCase(sorter.field),
       order_by: URL_SORT_ORDERS[sorter.order],
       page: pagination.current,
-      page_size: pagination.pageSize,
     };
 
     const pathname = window.location.pathname;