+++ stories/Components/Pagination.stories.jsx
@@ -1,5 +1,7 @@
 import React from "react";
 
+import { BrowserRouter } from "react-router-dom";
+
 import Pagination from "components/Pagination";
 
 import PaginationStoriesDocs from "!raw-loader!./PaginationStoriesDocs.mdx";
@@ -17,7 +19,11 @@ const metadata = {
   },
 };
 
-const Template = args => <Pagination {...args} />;
+const Template = args => (
+  <BrowserRouter>
+    <Pagination {...args} />
+  </BrowserRouter>
+);
 
 const Default = Template.bind({});
 Default.args = {